+++ apps/server/src/core/api/user/user.controller.ts
@@ -1,17 +1,15 @@
-import { Controller, Post, Body, BadRequestException, Get, Patch } from "@nestjs/common"
+import { Controller, Post, Body, BadRequestException, Get, Patch, Request, UseGuards } from "@nestjs/common"
 import { UserService } from "./user.service"
 import { GenerateAuthPasskeyDto } from "./dto/generate-auth-passkey.dto"
 import { VerifyAuthPasskeyDto } from "./dto/verify-auth-passkey.dto"
 import { statusMessages } from "src/utils/constants/status-messages"
-import { TokenAuthorizer, TokenAuthorizerResponse } from "src/auth/token-authorizer.decorator"
 import { UpdateCarbonSettingsDto } from "./dto/update-carbon-settings.dto"
-import { EventEmitter2 } from "@nestjs/event-emitter"
-import { EventsUnion } from "src/core/events/events.union"
 import { ChangeUsageInsightsSettingsDto } from "./dto/change-usage-insights.dto"
+import { ModRequest, TokenGuard } from "src/auth/token.guard"
 
 @Controller("user")
 export class UserController {
-  constructor(private readonly userService: UserService, private readonly eventEmitter: EventEmitter2) { }
+  constructor(private readonly userService: UserService) { }
 
   @Post("generatepasskey")
   async generateAuthPasskey(@Body() generateAuthPasskeyDto: GenerateAuthPasskeyDto) {
@@ -31,7 +29,6 @@ export class UserController {
       const response = await this.userService.verifyAuthPasskey(verifyAuthPasskeyDto)
 
       if (response.success) {
-        this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: response.user._id, module: "user", method: "POST", api: "/verifypasskey" })
         return { accessToken: response.accessToken, user: response.user }
       }
 
@@ -45,11 +42,11 @@ export class UserController {
     }
   }
 
+  @UseGuards(TokenGuard)
   @Get("userdetails")
-  async getUserDetails(@TokenAuthorizer() userT: TokenAuthorizerResponse) {
+  async getUserDetails(@Request() request: ModRequest) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: userT.userId, module: "user", method: "GET", api: "/userdetails" })
-      const { user, subscription, organization, hasActiveSubscription } = await this.userService.getUserDetails(userT.userId, userT.orgId)
+      const { user, subscription, organization, hasActiveSubscription } = await this.userService.getUserDetails(request.user.userId, request.user.orgId)
 
       if (user) {
         return { user, subscription, organization, hasActiveSubscription }
@@ -65,11 +62,11 @@ export class UserController {
     }
   }
 
+  @UseGuards(TokenGuard)
   @Post("signout")
-  async signOut(@TokenAuthorizer() user: TokenAuthorizerResponse) {
+  async signOut(@Request() request: ModRequest) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "user", method: "POST", api: "/signout" })
-      await this.userService.signOut(user.userId)
+      await this.userService.signOut(request.user.userId)
       return { message: statusMessages.signOutSuccess }
     }
 
@@ -78,11 +75,11 @@ export class UserController {
     }
   }
 
+  @UseGuards(TokenGuard)
   @Patch("updatecarbonsettings")
-  async updateCarbonSettings(@TokenAuthorizer() user: TokenAuthorizerResponse, @Body() updateCarbonSettingsDto: UpdateCarbonSettingsDto) {
+  async updateCarbonSettings(@Request() request: ModRequest, @Body() updateCarbonSettingsDto: UpdateCarbonSettingsDto) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "user", method: "PATCH", api: "/updatecarbonsettings" })
-      return await this.userService.updateCarbonSettings(user.userId, updateCarbonSettingsDto.reduceCarbonEmissions)
+      return await this.userService.updateCarbonSettings(request.user.userId, updateCarbonSettingsDto.reduceCarbonEmissions)
     }
 
     catch (error) {
@@ -90,11 +87,11 @@ export class UserController {
     }
   }
 
+  @UseGuards(TokenGuard)
   @Patch("changeusageinsights")
-  async changeUsageInsightsSettings(@TokenAuthorizer() user: TokenAuthorizerResponse, @Body() changeUsageInsightsSettingsDto: ChangeUsageInsightsSettingsDto) {
+  async changeUsageInsightsSettings(@Request() request: ModRequest, @Body() changeUsageInsightsSettingsDto: ChangeUsageInsightsSettingsDto) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "user", method: "PATCH", api: "/updatecarbonsettings" })
-      return await this.userService.changeUsageInsightsSettings(user.userId, changeUsageInsightsSettingsDto.usageInsights)
+      return await this.userService.changeUsageInsightsSettings(request.user.userId, changeUsageInsightsSettingsDto.usageInsights)
     }
 
     catch (error) {