+++ apps/server/src/core/events/insights/commands/impl/create-insights.command.ts
@@ -1,0 +1,7 @@
+import { CreateInsightsDto } from "../../dto/create-insights.dto"
+
+export class CreateInsightsCommand {
+  constructor(
+    public readonly createInsightsDto: CreateInsightsDto
+  ) { }
+}