+++ apps/server/src/core/events/insights/dto/create-insights.dto.ts
@@ -1,15 +1,12 @@
 import { IsNotEmpty } from "class-validator"
 
-export class CreateInsightDto {
+export class CreateInsightsDto {
   @IsNotEmpty()
   userId: string
-
-  @IsNotEmpty()
-  module: string
 
   @IsNotEmpty()
   method: string
 
   @IsNotEmpty()
-  api: string
+  apiUri: string
 }