+++ apps/server/src/products/httpnosql/httpnosql.controller.ts
@@ -2,17 +2,14 @@ import { Controller, Post, Body, Delete, Get, Patch, Param } from "@nestjs/commo
 import { HttpNosqlService } from "./httpnosql.service"
 import { CreateDataDto } from "./dto/create-data.dto"
 import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/auth/credential-authorizer.decorator"
-import { EventEmitter2 } from "@nestjs/event-emitter"
-import { EventsUnion } from "src/core/events/events.union"
 
 @Controller("products/httpnosql")
 export class HttpNosqlController {
-  constructor(private readonly httpNosqlService: HttpNosqlService, private readonly eventEmitter: EventEmitter2) { }
+  constructor(private readonly httpNosqlService: HttpNosqlService) { }
 
   @Post("create")
   async createKeyValue(@CredentialAuthorizer() user: CredentialAuthorizerResponse, @Body() createDataDto: CreateDataDto) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "products/httpnosql", method: "POST", api: "/create" })
       return await this.httpNosqlService.createKeyValue(user.orgId, createDataDto)
     }
 
@@ -24,7 +21,6 @@ export class HttpNosqlController {
   @Get("read")
   async readAllValues(@CredentialAuthorizer() user: CredentialAuthorizerResponse) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "products/httpnosql", method: "GET", api: "/read" })
       return await this.httpNosqlService.readAllValues(user.orgId)
     }
 
@@ -36,7 +32,6 @@ export class HttpNosqlController {
   @Get("read/:key")
   async readValueByKey(@CredentialAuthorizer() user: CredentialAuthorizerResponse, @Param() params: any) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "products/httpnosql", method: "GET", api: "/read" })
       return await this.httpNosqlService.readValueByKey(user.orgId, params.key)
     }
 
@@ -48,7 +43,6 @@ export class HttpNosqlController {
   @Patch("update")
   async updateValueByKey(@CredentialAuthorizer() user: CredentialAuthorizerResponse, @Body() updateDataDto: CreateDataDto) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "products/httpnosql", method: "DELETE", api: "/update" })
       return await this.httpNosqlService.updateValueByKey(user.orgId, updateDataDto)
     }
 
@@ -60,7 +54,6 @@ export class HttpNosqlController {
   @Delete("delete/:key")
   async deleteValueByKey(@CredentialAuthorizer() user: CredentialAuthorizerResponse, @Param() params: any) {
     try {
-      this.eventEmitter.emit(EventsUnion.CreateInsights, { userId: user.userId, module: "products/httpnosql", method: "DELETE", api: "/delete" })
       return await this.httpNosqlService.deleteValueByKey(user.orgId, params.key)
     }